+++ MapboxSearch/discover/src/main/java/com/mapbox/search/discover/Discover.kt
@@ -188,7 +188,7 @@ public interface Discover {
     }
 
     /**
-     * @suppress
+     * Companion object.
      */
     public companion object {